# generated-server/src/main/java/com/alumni/userservice/api/UsersApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.24.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.alumni.userservice.api;

import com.alumni.userservice.model.CreateUserRequest;
import com.alumni.userservice.model.ErrorResponse;
import org.springframework.lang.Nullable;
import java.util.UUID;
import com.alumni.userservice.model.UpdateUserRequest;
import com.alumni.userservice.model.User;
import com.alumni.userservice.model.UserPage;
import com.alumni.userservice.model.UserType;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-07-23T09:01:43.370133340Z[Africa/Dakar]", comments = "Generator version: 7.24.0")
@Validated
@Tag(name = "Users", description = "Gestion des profils utilisateurs")
public interface UsersApi {

    default UsersApiDelegate getDelegate() {
        return new UsersApiDelegate() {};
    }

    String PATH_CREATE_USER = "/users";
    /**
     * POST /users : Créer un profil utilisateur
     * Création d&#39;un profil métier après création de l&#39;utilisateur dans Keycloak. 
     *
     * @param createUserRequest  (required)
     * @return Profil créé (status code 201)
     *         or Requête invalide (status code 400)
     *         or Utilisateur déjà existant (status code 409)
     *         or Authentification requise ou jeton invalide (status code 401)
     *         or AccÃ¨s refusÃ© (status code 403)
     */
    @Operation(
        operationId = "createUser",
        summary = "Créer un profil utilisateur",
        description = "Création d'un profil métier après création de l'utilisateur dans Keycloak. ",
        tags = { "Users" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Profil créé", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = User.class))
            }),
            @ApiResponse(responseCode = "400", description = "Requête invalide", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "409", description = "Utilisateur déjà existant", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Authentification requise ou jeton invalide", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "403", description = "AccÃ¨s refusÃ©", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = UsersApi.PATH_CREATE_USER,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<User> createUser(
        @Parameter(name = "CreateUserRequest", description = "", required = true) @Valid @RequestBody CreateUserRequest createUserRequest
    ) {
        return getDelegate().createUser(createUserRequest);
    }


    String PATH_DELETE_USER = "/users/{userId}";
    /**
     * DELETE /users/{userId} : Supprimer un utilisateur
     *
     * @param userId  (required)
     * @return Utilisateur supprimé (status code 204)
     *         or Requête invalide (status code 400)
     *         or Ressource introuvable (status code 404)
     *         or Authentification requise ou jeton invalide (status code 401)
     *         or AccÃ¨s refusÃ© (status code 403)
     */
    @Operation(
        operationId = "deleteUser",
        summary = "Supprimer un utilisateur",
        tags = { "Users" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Utilisateur supprimé"),
            @ApiResponse(responseCode = "400", description = "Requête invalide", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Ressource introuvable", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Authentification requise ou jeton invalide", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "403", description = "AccÃ¨s refusÃ©", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = UsersApi.PATH_DELETE_USER,
        produces = { "application/json" }
    )
    default ResponseEntity<Void> deleteUser(
        @Parameter(name = "userId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("userId") UUID userId
    ) {
        return getDelegate().deleteUser(userId);
    }


    String PATH_GET_CURRENT_USER = "/users/me";
    /**
     * GET /users/me : Récupérer le profil de l&#39;utilisateur connecté
     *
     * @return Profil courant (status code 200)
     *         or Authentification requise ou jeton invalide (status code 401)
     *         or AccÃ¨s refusÃ© (status code 403)
     */
    @Operation(
        operationId = "getCurrentUser",
        summary = "Récupérer le profil de l'utilisateur connecté",
        tags = { "Users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Profil courant", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = User.class))
            }),
            @ApiResponse(responseCode = "401", description = "Authentification requise ou jeton invalide", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "403", description = "AccÃ¨s refusÃ©", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = UsersApi.PATH_GET_CURRENT_USER,
        produces = { "application/json" }
    )
    default ResponseEntity<User> getCurrentUser(
        
    ) {
        return getDelegate().getCurrentUser();
    }


    String PATH_GET_USER_BY_ID = "/users/{userId}";
    /**
     * GET /users/{userId} : Obtenir un utilisateur par ID
     *
     * @param userId  (required)
     * @return Utilisateur trouvé (status code 200)
     *         or Requête invalide (status code 400)
     *         or Ressource introuvable (status code 404)
     *         or Authentification requise ou jeton invalide (status code 401)
     *         or AccÃ¨s refusÃ© (status code 403)
     */
    @Operation(
        operationId = "getUserById",
        summary = "Obtenir un utilisateur par ID",
        tags = { "Users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Utilisateur trouvé", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = User.class))
            }),
            @ApiResponse(responseCode = "400", description = "Requête invalide", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Ressource introuvable", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Authentification requise ou jeton invalide", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "403", description = "AccÃ¨s refusÃ©", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = UsersApi.PATH_GET_USER_BY_ID,
        produces = { "application/json" }
    )
    default ResponseEntity<User> getUserById(
        @Parameter(name = "userId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("userId") UUID userId
    ) {
        return getDelegate().getUserById(userId);
    }


    String PATH_SEARCH_USERS = "/users";
    /**
     * GET /users : Rechercher les utilisateurs
     *
     * @param name  (optional)
     * @param email  (optional)
     * @param type  (optional)
     * @param page  (optional, default to 0)
     * @param size  (optional, default to 20)
     * @return Liste paginée des utilisateurs (status code 200)
     *         or Requête invalide (status code 400)
     *         or Authentification requise ou jeton invalide (status code 401)
     *         or AccÃ¨s refusÃ© (status code 403)
     */
    @Operation(
        operationId = "searchUsers",
        summary = "Rechercher les utilisateurs",
        tags = { "Users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Liste paginée des utilisateurs", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserPage.class))
            }),
            @ApiResponse(responseCode = "400", description = "Requête invalide", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Authentification requise ou jeton invalide", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "403", description = "AccÃ¨s refusÃ©", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = UsersApi.PATH_SEARCH_USERS,
        produces = { "application/json" }
    )
    default ResponseEntity<UserPage> searchUsers(
        @Size(min = 1, max = 200) @Parameter(name = "name", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "name", required = false) @Nullable String name,
        @jakarta.validation.constraints.Email @Parameter(name = "email", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "email", required = false) @Nullable String email,
        @Parameter(name = "type", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "type", required = false) @Nullable UserType type,
        @Min(value = 0) @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Integer page,
        @Min(value = 1) @Max(value = 100) @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "20") Integer size
    ) {
        return getDelegate().searchUsers(name, email, type, page, size);
    }


    String PATH_UPDATE_CURRENT_USER = "/users/me";
    /**
     * PATCH /users/me : Modifier son propre profil
     *
     * @param updateUserRequest  (required)
     * @return Profil mis à jour (status code 200)
     *         or Requête invalide (status code 400)
     *         or Authentification requise ou jeton invalide (status code 401)
     *         or AccÃ¨s refusÃ© (status code 403)
     */
    @Operation(
        operationId = "updateCurrentUser",
        summary = "Modifier son propre profil",
        tags = { "Users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Profil mis à jour", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = User.class))
            }),
            @ApiResponse(responseCode = "400", description = "Requête invalide", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Authentification requise ou jeton invalide", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "403", description = "AccÃ¨s refusÃ©", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PATCH,
        value = UsersApi.PATH_UPDATE_CURRENT_USER,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<User> updateCurrentUser(
        @Parameter(name = "UpdateUserRequest", description = "", required = true) @Valid @RequestBody UpdateUserRequest updateUserRequest
    ) {
        return getDelegate().updateCurrentUser(updateUserRequest);
    }


    String PATH_UPDATE_USER = "/users/{userId}";
    /**
     * PATCH /users/{userId} : Modifier un utilisateur
     *
     * @param userId  (required)
     * @param updateUserRequest  (required)
     * @return Utilisateur modifié (status code 200)
     *         or Requête invalide (status code 400)
     *         or Ressource introuvable (status code 404)
     *         or Authentification requise ou jeton invalide (status code 401)
     *         or AccÃ¨s refusÃ© (status code 403)
     *         or Conflit avec une ressource existante (status code 409)
     */
    @Operation(
        operationId = "updateUser",
        summary = "Modifier un utilisateur",
        tags = { "Users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Utilisateur modifié", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = User.class))
            }),
            @ApiResponse(responseCode = "400", description = "Requête invalide", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Ressource introuvable", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Authentification requise ou jeton invalide", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "403", description = "AccÃ¨s refusÃ©", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "409", description = "Conflit avec une ressource existante", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PATCH,
        value = UsersApi.PATH_UPDATE_USER,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<User> updateUser(
        @Parameter(name = "userId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("userId") UUID userId,
        @Parameter(name = "UpdateUserRequest", description = "", required = true) @Valid @RequestBody UpdateUserRequest updateUserRequest
    ) {
        return getDelegate().updateUser(userId, updateUserRequest);
    }

}